net: ipv4: Add a sysctl to set multipath hash seed
When calculating hashes for the purpose of multipath forwarding, both IPv4
and IPv6 code currently fall back on flow_hash_from_keys(). That uses a
randomly-generated seed. That's a fine choice by default, but unfortunately
some deployments may need a tighter control over the seed used.

In this patch, make the seed configurable by adding a new sysctl key,
net.ipv4.fib_multipath_hash_seed to control the seed. This seed is used
specifically for multipath forwarding and not for the other concerns that
flow_hash_from_keys() is used for, such as queue selection. Expose the knob
as sysctl because other such settings, such as headers to hash, are also
handled that way.

Despite being placed in the ipv4 namespace, the multipath seed sysctl is
used for both IPv4 and IPv6, similarly to e.g. a number of TCP variables.

The seed used by flow_hash_from_keys() is a 128-bit quantity. However it
seems that usually the seed is a much more modest value. 32 bits seem
typical (Cisco, Cumulus), some systems go even lower. Therefore go with a
32-bit quantity, which is then quadruplicated to form the siphash key.

In order to keep the existing flow hash seed secret, maintain the siphash
separately from the user-set seed. This allows setting the siphash to the
value of the global flow hash seed by default, and override when the user
sets it. At the same time, separately from the siphash, the user can read
and write "their" 32-bit seed, which then controls the siphash value.

Signed-off-by: Petr Machata <petrm@nvidia.com>