net: flow_dissector: Always initialize the hash key
The following patches will make it possible for a user to configure the
hash seed used for multipath hash forwarding. The idea is that there is
always a dedicated seed, which multipath forwarding uses when computing
flow hash. Until the user sets the value, the value of this seed is the
same as the secret hash used by the flow dissector. Then after the user
sets the seed, a new hash key is generated.

For this to work, the sysctl code, as it constructs a netns, will query the
value of the secret seed. Therefore the secret seed will have to be
constructed as soon as an init_net is constructed. The late initialization
is therefore moot and the value can just be constructed during init.

To that end, convert init_default_flow_dissectors() into just a helper, and
add a new function, flow_dissector_init(), to hold any initialization in
general. Have that function call init_default_flow_dissectors(), and also
__flow_hash_secret_init().

Signed-off-by: Petr Machata <petrm@nvidia.com>