selftests: forwarding: mirror_gre_lag_lacp: Simplify
This tests gretap mirroring to a team netdevice. The test runs twice --
once in skip_hw mode, testing the SW datapath, once in skip_sw mode to test
the HW datapath. But in order to test the SW datapath on a HW topology, it
needs to resort to trapping traffic that would otherwise be HW-forwarded.
This unnecessarily complicates the test -- the simple solution to test the
SW datapath is to just run the selftest with veth pairs.

So here, just drop the complexity around the trap installation.

Signed-off-by: Petr Machata <petrm@nvidia.com>