selftests: forwarding: mirror_gre_lag_lacp: Make less noisy
The test works by sending packets through a tunnel, whence they are
forwarded to a LAG. One of the LAG children is removed from the LAG prior
to the exercise, and the test then counts how many packets pass through the
other one. The issue with this is that it counts all packets, not just the
encapsulated ones. At the same time, adding a sophisticated filter might be
problematic for offloaded datapaths.

So instead add a second gretap endpoint to receive the sent packets, and
check reception counters there.

Furthermore, refine the mirroring condition so that only ICMP packets get
mirrored, and counted.

With these measures in place, the test looks robust even on my laptop,
where it would previously fail more often than not.

Signed-off-by: Petr Machata <petrm@nvidia.com>