selftests: forwarding: mirror_lib: Drop one check
The check sends ICMP traffic in a "backward" direction, and expects to see
ICMP responses mirrored. This will not work when we move to UDP traffic in
the following patches. The check also is not particularly interesting,
because mirroring as such is verified in the previous check. Drop this
check.

Signed-off-by: Petr Machata <petrm@nvidia.com>