selftests: mirror libs: Parameterize which host sends traffic
Currently mirroring tests always send traffic from v$h1. If the opposite
direction is to be tested, it is done through the ICMP responses that the
sent traffic elicits. In the future, the mirroring tests should be using
UDP traffic to make them less noisy, and this trick does not work anymore.
It will be necessary to actually send the traffic in the opposite
direction. The IP addresses are already configurable, in this patch make
the sending host configurable as well.

Signed-off-by: Petr Machata <petrm@nvidia.com>