selftests: mirror: Use UDP for testing traffic
Mirroring tests used to just use ping to generate testing traffic. Because
this led to issues with stability, eventually they were migrated over to
mausezahn, however still kept using ICMP and ICMPv6. This still causes
stability issues, because when checking for number of packets, the probes
get confused by e.g. router solicitation messages. In this patch, do what
we should have done from the beginning: use UDP with unique port numbers.

This is straightforward as long as the mirroring is done on an ingress of
an interface. Egress testing however was done by sending traffic in the
original direction, which caused the recipient to echo back, which was
mirrored. Instead, in this patch, simply send traffic in the opposite
direction. This patch adds a number of _rev_ functions that do what their
non-rev counterparts do, except in the reverse direction.

Note that at this point, the names icmp_capture_install() et.al. are
misnomers. This will be fixed shortly.

Signed-off-by: Petr Machata <petrm@nvidia.com>