selftests: mirror: Demand accurate number of packets in most tests
Now that UDP is used, there's no reason not to just demand that exactly the
number of packets sent be received on the opposite end of the tunnel.
Except that's not exactly right: some tests put probes at netdevices that
carry encapsulated traffic. In matchall mirrors, those will see traffic
including all sorts of noise, and because it's encapsulated, it won't be
possible to check that it's our UDP.

Therefore change mirror_test() to either take an exact number of packets to
check for, or an expression. In the several places where non-exact match is
necessary, simply use an expression involving >= to check for at least the
desired number of packets.

Signed-off-by: Petr Machata <petrm@nvidia.com>