net: nexthop: Add flag to assert that NHGRP reserved fields are zero [xxx][old]
There are many unpatched kernel versions out there that do not initialize
the reserved fields of struct nexthop_grp. The issue with that is that if
those fields were to be used for some end (i.e. stop being reserved), old
kernels would still keep sending random data through the field, and a new
userspace could not rely on the value.

In this patch, add a new flag to nexthop messages, which indicates that the
reserved fields in struct nexthop_grp are zeroed before dumping. This is
reliant on the actual fix from commit [xxx] ("net: nexthop: Initialize all
fields in dumped nexthops").

Signed-off-by: Petr Machata <petrm@nvidia.com>