mlxsw: pci: Initialize XDP Rx queue info per port
'struct xdp_rxq_info' was added to provide more info for XDP program.
The idea is to initialize the structure as part of init and attach it to
XDP buffer, so instead of updating N-members in xdp_buff, we can use the
same pre-initialized info. Rx info registration requires:
* netdevice
* Queue index
* NAPI ID

To pre-initialize such structure with all the info we need
{netdev} x {queue} pairs, as there is no queue for each netdevice using
Spectrum ASICs. This solution is optional, but creates huge overhead.

One option is to setup such structure per Rx queue, and use dummy
netdevice, 'xdp_rxq_info->dev' is used to set 'xdp_md->ingress_ifindex',
which means that using dummy netdev, XDP program will not be aware of the
Rx port.

Another option is to store such structure per port, as the Rx port is
useful info for XDP program and we would like to set the actual 'ifindex'.
That means that queue index and NAPI ID are not set. It's ok to not set
queue index, the meaning is that XDP program will not have this info.
Regarding NAPI ID, it is used to find the corresponding NAPI structure
for socket busy polling, which is relevant for XSK, which is now not
planned to be supported in mlxsw.

Set XDP Rx queue info per port, to provide 'ifindex' for XDP program, as
for now it looks like the most interesting field. In future we can
create many pairs of {netdev} x {queue}, so all the fields will be set
correctly. Handle Rx queue info when netdevice is set and unset.

Signed-off-by: Amit Cohen <amcohen@nvidia.com>
Reviewed-by: Ido Schimmel <idosch@nvidia.com>