net: fib_nexthops: Dump 16-bit nexthop weights
In the commit cited below, (n)ECMP nexthop weight was increased from 8 bits
to 16 bits. The patchset however missed that "ip route" dumps the weights
as well, even for nexthop groups. Thus:

 # ip link add name a up type dummy
 # ip nexthop add id 101 dev a
 # ip nexthop add id 102 dev a
 # ip nexthop add id 103 dev a
 # ip nexthop add id 104 dev a
 # ip nexthop add id 105 dev a
 # ip nexthop add id 100 group 101,1/102,10/103,100/104,1000/105,10000
 # ip route add 192.0.0.0/28 nhid 100
 # ip route show 192.0.0.0/28192.0.0.0/28 nhid 100
        nexthop dev a weight 1
        nexthop dev a weight 10
        nexthop dev a weight 100
        nexthop dev a weight 232
        nexthop dev a weight 16

struct rtnexthop has no free bits anymore where the extra bit could be
stored. So in this patch, add a new attribute, and if the weight does not
fit into 8 bits, attach it as an attribute.

The question is what to do with the existing 8-bit value in that case.
Unlike with struct nexthop_grp, when the attribute is attached, the field
is unused.

This patch opts to set it to 0xff. This value was always theoretically
allowed, to represent a nexthop with weight of 256. But attempts to
actually configure such next hops were bounced. Thus it can serve as an
out-of-band value to indicate something is up.

At the same time, 256 is closer to the true weight of the next hop than
weight modulo 256.

Fixes: b72a6a7ab957 ("net: nexthop: Increase weight to u16")
Signed-off-by: Petr Machata <petrm@nvidia.com>