selftests: net: tcpdump_stop(): Don't kill via ip netns [xxx]
[xxx] test me

tcpdump_stop() stops a process previously launched by tcpdump_start(). In
tcpdump_start(), the process is sometimes started through "ip netns exec"
to bind the tcpdump socket correctly. tcpdump_stop() mimics this behavior,
however unlike PID, it does not carrying the NS-exec command through a
helper array, relying instead on just picking up the last set global.
But it is not necessary to kill the process in a particular network
namespace, so none of this is necessary.

Instead, make ns_cmd local in tcpdump_start(), and then do not use it in
tcpdump_stop(). Instead, just invoke the existing stop_process() to carry
out the process termination. This takes care of any termination noise, too.

Signed-off-by: Petr Machata <petrm@nvidia.com>