selftests: forwarding: lib: Move packet-gen helpers to a new library
A number of functions in forwarding/lib.sh deal with generating payloads
for mausezahn. These functions are not necessarily forwarding-related, and
could be in a separate library that interested parties source as needed
from anywhere.

Move these functions into a new dedicated library, pkt.sh, and have current
users include the library explicitly.

Signed-off-by: Petr Machata <petrm@nvidia.com>