selftests: net: lib: Add a new helper, autodefer()
In order to simplify scheduling cleanups, add a helper, which takes a
template on input, and out of that, deduces how both the setup and the
teardown legs of a command should look like. Then it issues the setup right
away and defer the cleanup.

The operation is currently very simple: each word of the template is either
taken as is and added to both setup and cleanup commands, or is split along
a /// separator, where the front part is added to setup and the back part
to cleanup.

For example:

	autodefer ip link set dev $swp1 up///down

Signed-off-by: Petr Machata <petrm@nvidia.com>