net: vxlan: Allow separation of RX and TX interfaces for MC traffic [xxx]
When the remote address of a VXLAN tunnel is a multicast group, two
interfaces come in play, conceptually:

- the one through which the traffic is sent, and is therefore used as an
  iif in MC routing table
- the one through which the traffic is received, and is therefore the one
  where IGMP membership report should be issued

[xxx] Still unclear why exactly it is necessary.

The two are currently always the same interface, the one given as
IFLA_VXLAN_LINK. Some installations however use two different interfaces
for the purpose. Instead of introducing another link argument, this patch
allows lookup of the TX interface by the local endpoint address.

Add a new flag, IFLA_VXLAN_MC_NODEV, that causes OIF of outgoing packets to
be set to 0, therefore prompting a lookup of a local route matching local
endpoint address, and deducing the OIF to use from that.

This patch is based on an earlier patch from Nikolay Aleksandrov and Roopa
Prabhu, though it underwent significant changes.

Cc: Nikolay Aleksandrov <razor@blackwall.org>
Cc: Roopa Prabhu <roopa@nvidia.com>
Signed-off-by: Petr Machata <petrm@nvidia.com>