fib6_is_reject += IPV6_ADDR_MULTICAST
Today, on a loopback netdevice, any route other than a loopback route on is
marked as RTF_REJECT. This patch prevents this from happening for multicast
routes. It is necessary for oif=0 lookup to work on IPv6.

An additional requirement is that a multicast route must exist for the
loopback netdevice (which would normally be marked as reject, but with this
patch it won't). I think (untested) that it has to be more specific than a
ff00::/8 route, because many netdevices have that, and it's not guaranteed
that `lo' would be picked over these.