Broken IPv6 nodev tests
These don't work or work by accident. IPv6 matches on the MC route and
doesn't do netdevice lookup by source address. Which of the several
ff00::/8 routes matches is unpredictable. It just happens to be lo10 like
we need to in the "normal" case, but fails where we want the system to go
through lo instead of lo10.