selftests: forwarding: lib: Split setup_wait()
setup_wait() takes an optional argument and then is called from the top
level of the test script. That confuses shellcheck, which think maybe the
intention is to pass $1 of the script to the function, which is never the
case. To avoid having to annotate every single new test with a SC disable,
split the function to two: one that takes a mandatory argument, and one
that takes no argument at all.

Convert the two existing users of that optional argument, both in Spectrum
resource selftest, to use the new form. Clean up vxlan_bridge_1q_mc_ul.sh
to not pass a now-unused argument.

Signed-off-by: Petr Machata <petrm@nvidia.com>