net: bridge: Use red-black tree for FDB dumps
As the number of FDB entries grows, the time to dump FDB becomes dominated
by the time spent skipping already-dumped entries, and the dump gets
disproportionately expensive. For FDB sized at 150K entries, the dump times
already hit a second, which is pretty rough.

So one idea is to organize the FDB entries in a data structure more
amenable to restarts. Currently we just keep all entries on a single linked
list.

One option is to keep the entries in a red-black tree, which is the
approach presented here. The entries are sorted by the hash key, i.e.
the (ETH address, VID) tuple. This helps quite a bit with the dumping. In
experiments that I did with 400K-entry FDB, it cut down the time it took to
serialize the FDB by some 60 % (from 6.5 s to 2.5 s).

However, there are two downsides:

- this makes the insertion more expensive. The tree needs to be rebalanced
  under lock. Maybe this is not such a big deal, because for every FDB
  entry, learning happens once, and then is likely reused many times. But
  it does make things slower. I can measure it if necessary.

- dumping needs to be done under spin lock. Because all the rotations
  involved in insertion, I suspect the tree will not be kept walkable
  throughout. The documentation demands locking in any case. That means an
  ongoing dump will hold learning up. I feel like this is the true killer.

The above two points are the reason I'm sending the patchset -- I'd like
some feedback about whether this seems to make sense to anyone. Or maybe
y'all see some obvious way out.

- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -

As an appendix, I should explain how the patch uses the ndo_fdb_dump's
index parameter, and why I consider it unwise:

Because bridges can't be bridge slaves, bridge's ndo_fdb_dump will be
invoked at most once from rtnl_fdb_dump(), and if more than one
ndo_fdb_dump operation will take place, the bridge one will be first. Since
all other NDO implementations use a simple linear index allocation, we have
a relative freedom in how we allocate indices in the bridge driver. The any
other NDO will just continue where-ever the bridge leaves off. All Ethernet
addresses can be used (it is legal to add MC/BC addresses to the FDB as
well), but VID only takes 12 bits. That gives us 4 bits of guaranteed
zeroes after the bridge FDB iteration is over, or 16 times as much index
space for other dump implementors. So far so good, this should work.

The reason I do not think it is a good idea anymore is that the scheme is
fairly complex. There are several non-obvious conditions that keep it
afloat. The way that rtnl_fdb_dump() is structured, how others allocate
indices, how others _compare_ indices. Currently the code assumes that it
owns the start of the index space, which is not in the API contract. I
think a proper solution would use augmented RB tree to maintain a count of
all nodes in a given subtree, and use that to skip appropriately. We get
count as input parameter, and return count dumped as output, and this would
respect that API.

Signed-off-by: Petr Machata <petrm@nvidia.com>